Write sizeof obj.
The c++ sizeof operater returns the size of its argument. We want to
return the size of the object, in case the size varies. If the object is
a string, the size varies with the string. If you don't want the size to
vary, use an object that always has the same size.